Added new functions for names and employee number
Added functions for employee names and numbers and implemented this into exportData(). Problems in the find functions are causing list lengths to vary and as such writes cannot be completed. Since the requirements of this are pretty sandboxed I will go for the simplest solution in the future. Current counts for reference:
175
80
70
116
35
54
1090
1185