Added testing files
Added the necessary workbooks needed for testing as well as a comparexltesting.py file to use with these. The file names are already configured and ready to run. Simply download these files and run the testing script when everything is in the same directory. Only the five files from this commit are required for testing